requirements.txt
fastapi
uvicorn 
openai 
supabase
streamlit
requests
dotenv
psycopg2-binary
supervisor